Refactor julia initialization by finding and loading the sysimage earlier (#58231)
This changes the order of initialization slightly. The goal is to be
able to initialize executables and share libraries without having to
explicitly call init functions.

This still doesn't address the thread safety
