+++ test/trimming/init.c
@@ -1,9 +1,11 @@
+#define _GNU_SOURCE
+#include <dlfcn.h>
 #include <julia.h>
 
 __attribute__((constructor)) void static_init(void)
 {
     if (jl_is_initialized())
         return;
-    julia_init(JL_IMAGE_IN_MEMORY);
+    jl_init_with_image_handle((void *)RTLD_DEFAULT);
     jl_exception_clear();
 }
